requirements.txt
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-CeleryExt==0.5.0
redis==7.0.1
flask-wtf==1.2.2
flask-debugtoolbar==0.16.0
flask-restplus==0.13.0
bootstrap-flask~=2.5
beautifulsoup4==4.14.2
celery==5.5.3
pyasn==1.6.2
requests==2.32.5
dnspython==2.8.0
python-strtobool==1.0.3
lxml==6.0.2
flask-login==0.6.3
tldextract==5.3.0
cryptography==46.0.3
celery-redbeat==2.3.3 
pymysql==1.1.2
flask_restx==1.3.2
marshmallow-sqlalchemy==1.4.2
pytest==8.4.2
pytest-cov==7.0.0
coverage==7.11.0
flask-marshmallow==1.3.0
gunicorn==23.0.0
